Do some unit test for make sure components are create

+++ src/app/components/product-checkout/product-checkout.component.ts
@@ -1,17 +1,20 @@
 import { Component, OnInit, Input } from '@angular/core';
 import { product } from '../../services/products.service';
-import { cart, cartService } from 'src/app/services/cart.service';
+import { cartService } from 'src/app/services/cart.service';
 import { faTrash } from '@fortawesome/free-solid-svg-icons';
+
 @Component({
   selector: 'app-product-checkout',
   templateUrl: './product-checkout.component.html',
   styleUrls: ['./product-checkout.component.css'],
 })
 export class ProductCheckoutComponent implements OnInit {
-  @Input() product: product | any;
+  @Input() product: product | undefined;
   faTrash = faTrash;
   removeProductFromCart() {
-    this.cartService.removeProductFromCart(this.product);
+    if (this.product) {
+      this.cartService.removeProductFromCart(this.product);
+    }
   }
   constructor(private cartService: cartService) {}
